requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
openai==1.5.0
pandas==2.1.0
scikit-learn==1.3.0
numpy==1.25.2
pytest==7.4.0
SQLAlchemy==2.0.20
aiohttp==3.8.5 